FEATURE: rename MovieDetailsPage.tsx to MovieDetails.tsx, and update main.tsx based on rename
